requirements.txt
pycryptodome==3.17
# Security and Connectivity
requests[security]>=2.29.0
